requirements.txt
flask
flask-cors
python-dotenv
huggingface_hub
requests
